+++ packages/bundle-size/src/plugin-blockquote.tsx
@@ -1,0 +1,1 @@
+export { blockquotePlugin } from '@edtr-io/plugin-blockquote'