+++ packages/bundle-size/src/plugin-highlight.tsx
@@ -1,0 +1,1 @@
+export { highlightPlugin } from '@edtr-io/plugin-highlight'