+++ packages/bundle-size/src/plugin-input-exercise.tsx
@@ -1,0 +1,1 @@
+export { inputExercisePlugin } from '@edtr-io/plugin-input-exercise'