+++ packages/core/src/store/focus/reducer.ts
@@ -149,6 +149,7 @@ export interface Node {
 
 export const publicFocusSelectors = {
   isFocused,
+  getFocused,
   getFocusTree,
   findPreviousNode,
   findNextNode