+++ packages/plugin-rows/src/editor/index.tsx
@@ -1,4 +1,9 @@
-import { StatefulPluginEditorProps, selectors, useStore } from '@edtr-io/core'
+import {
+  StatefulPluginEditorProps,
+  selectors,
+  useStore,
+  ScopeContext
+} from '@edtr-io/core'
 import * as React from 'react'
 
 import { rowsState } from '..'
@@ -9,12 +14,11 @@ export const RowsEditor = (
 ) => {
   const rows = props.state
   const store = useStore()
-
+  const { scope } = React.useContext(ScopeContext)
   return (
     <React.Fragment>
       {rows.items.map((row, index) => {
         const doc = selectors.getDocument(store.getState(), row.id)
-        const plugins = selectors.getPlugins(store.getState())
 
         if (!doc) return null
 
@@ -24,10 +28,10 @@ export const RowsEditor = (
               {...props}
               index={index}
               doc={doc}
-              store={store}
+              fullStore={store}
               moveRow={rows.move}
               insert={rows.insert}
-              plugins={plugins}
+              scope={scope}
             />
           </div>
         )