+++ scripts/changelog.ts
@@ -299,7 +299,12 @@ async function exec(): Promise<void> {
       ]
     },
     {
+      changed: [
+        'Optimize bundle size',
+        '**plugin-rows**. add-button is now always visible under the focused document'
+      ],
       added: [
+        '**plugin-serlo-injection**. Added injection plugin for serlo.org content',
         '**plugin-table**. You can now create tables in markdown.',
         '**plugin-important-statement**. Plugin that highlights important statements.'
       ]